added files to be excluded from rat license requirement, abstracted out the common code in both predict and train classes into NaiveBayesHelper class that now has a generateLabeledPoints method which use to be getData. The new method name is descriptive
